thinking about how to implement Attend
