+++ src/app/dbase/data/data.service.ts
@@ -5,9 +5,7 @@ import { tap } from 'rxjs/operators';
 import { Store } from '@ngxs/store';
 import { SLICE } from '@dbase/state/store.define';
 
-import {
-  filterTable
-} from '@dbase/app/app.library';
+import { filterTable } from '@dbase/app/app.library';
 import { COLLECTION, FIELD, FILTER } from '@dbase/data/data.define';
 import { IStoreMeta } from '@dbase/data/data.interface';
 import { IWhere } from '@dbase/fire/fire.interface';
@@ -15,14 +13,13 @@ import { FireService } from '@dbase/fire/fire.service';
 import { SyncService } from '@dbase/sync/sync.service';
 import { AuthService } from '@dbase/auth/auth.service';
 
-import { cryptoHash } from '@lib/crypto.library';
 import { getStamp } from '@lib/date.library';
 import { IObject } from '@lib/object.library';
 import { dbg } from '@lib/logger.library';
 
 /**
- * DataService is responsible for managing the syncing between the local State (ngxs 'Store')
- * and the remote DataBase ('DBaseService').  
+ * DataService is responsible for managing the syncing between
+ * the local State (ngxs 'Store') and the remote DataBase (Firebase Firestore).  
  * The intention is that all 'reads' are from State, and all 'writes' are to a persisted
  * local copy of the Database (which FireStore will sync back to the server).
  */
