+++ src/app/route/profile/plan/plan.component.ts
@@ -3,35 +3,17 @@ import { Observable } from 'rxjs';
 import { Select } from '@ngxs/store';
 import { ClientState } from '@dbase/state/client.state';
 
-import { IProfilePlan, IPrice } from '@dbase/data/data.interface';
-import { COLLECTION, FIELD, STORE } from '@dbase/data/data.define';
-import { DataService } from '@dbase/data/data.service';
-import { dbg } from '@lib/logger.library';
+import { MemberService } from '@dbase/app/member.service';
+import { IPrice } from '@dbase/data/data.interface';
 
 @Component({
   selector: 'wb-plan',
   templateUrl: './plan.component.html',
 })
 export class PlanComponent implements OnInit {
-  @Select(ClientState.plans) plan$!: Observable<IProfilePlan[]>;
+  @Select(ClientState.prices) price$!: Observable<IPrice[]>;
 
-  private dbg: Function = dbg.bind(this);
-
-  constructor(private readonly data: DataService) { }
+  constructor(private readonly member: MemberService) { }
 
   ngOnInit() { }
-
-  setPlan(price: IPrice) {
-    this.dbg('price: %j', price);
-    const planDoc: IProfilePlan = {
-      [FIELD.id]: this.data.newId,                    // generate a new Id
-      store: STORE.profile,
-      type: 'plan',
-      plan: price.plan,
-    }
-
-    this.dbg('plan: %j', planDoc);
-    this.data.insDoc(COLLECTION.Member, planDoc)
-      .catch(err => this.dbg('setPlan: %j', err))
-  }
 }
