+++ src/app/route/routing.module.ts
@@ -1,17 +1,17 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-
-import { MaterialModule } from '@app/route/material.module';
 import { Routes, RouterModule } from '@angular/router';
-import { AuthGuard, ProfileGuard } from '@route/routing.guard';
 
+import { MaterialModule } from '@route/material.module';
 import { LoginComponent } from '@route/login/login.component';
 import { AttendComponent } from '@route/attend/attend.component';
+import { AuthGuard, ProfileGuard } from '@route/routing.guard';
 
 const routes: Routes = [
 	{ path: 'login', component: LoginComponent },
 	{ path: 'attend', component: AttendComponent, canActivate: [AuthGuard, ProfileGuard] },
 	{ path: 'profile', loadChildren: '@route/profile/profile.module#ProfileModule' },
+	{ path: 'admin', loadChildren: '@route/admin/admin.module#AdminModule' },
 	{ path: '**', redirectTo: '/attend', pathMatch: 'full' },
 ];
